Added login/logout check for schemas

+++ schemas/views.py
@@ -11,63 +11,77 @@ from .forms import SchemaForm, ColumnForm
 
 
 def DashboardView(request):
+    if request.user.is_authenticated:
 
-    schemas = Schema.objects.all()
+        schemas = Schema.objects.all()
 
-    context = {
+        context = {
 
-        'schemas': schemas
-    }
-    return render(request=request, template_name="schemas/dashboard.html", context=context)
+            'schemas': schemas
+        }
+        return render(request=request, template_name="schemas/dashboard.html", context=context)
+
+    else:
+        return HttpResponseRedirect('login/')
 
 
 def NewSchemaView(request):
+    if request.user.is_authenticated:
 
-    if request.method == "POST":
+        if request.method == "POST":
 
-        form = SchemaForm(request.POST)
+            form = SchemaForm(request.POST)
+
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect("/")
+
+        else:
+            form = SchemaForm()
+
+        context = {'form': form}
+
+        return render(request=request, template_name="schemas/schema.html", context=context)
+
+    else:
+        return HttpResponseRedirect("login/")
+
+
+def DeleteSchemaView(request, schema_id):
+    if request.user.is_authenticated:
+
+        try:
+            schema = Schema.objects.get(pk=schema_id)
+
+        except Schema.DoesNotExist:
+
+            return HttpResponseRedirect("/")
+
+        schema.delete()
+        return HttpResponseRedirect("/")
+
+    else:
+        return HttpResponseRedirect("/login/")
+
+
+def UpdateSchemaView(request, schema_id):
+    if request.user.is_authenticated:
+
+        schema = Schema.objects.get(pk=schema_id)
+        schema.last_modified = datetime.datetime.now()
+
+        form = SchemaForm(request.POST or None, instance=schema)
 
         if form.is_valid():
             form.save()
             return HttpResponseRedirect("/")
 
+        context = {'form': form}
+
+        return render(request=request, template_name="schemas/schema.html", context=context)
+
     else:
-        form = SchemaForm()
-
-    context = {'form': form}
-
-    return render(request=request, template_name="schemas/schema.html", context=context)
-
-
-# def DeleteColumnView(request):
-
-def DeleteSchemaView(request, schema_id):
-
-    try:
-        schema = Schema.objects.get(pk=schema_id)
-
-    except Schema.DoesNotExist:
-
-        return HttpResponseRedirect("/")
-
-    schema.delete()
-    return HttpResponseRedirect("/")
-
-
-def UpdateSchemaView(request, schema_id):
-
-    schema = Schema.objects.get(pk=schema_id)
-    schema.last_modified = datetime.datetime.now()
-
-    form = SchemaForm(request.POST or None, instance=schema)
-
-    if form.is_valid():
-        form.save()
-        return HttpResponseRedirect("/")
-
-    context = {'form': form}
-
-    return render(request=request, template_name="schemas/schema.html", context=context)
+        return HttpResponseRedirect("/login/")
 
 
 def ColumnsDashboardView(request, schema_id):
@@ -104,8 +118,6 @@ def NewColumnView(request, schema_id):
     return render(request=request, template_name="schemas/column.html", context=context)
 
 
-# def DeleteColumnView(request):
-
 def DeleteColumnView(request, column_id, schema_id):
 
     try:
@@ -135,45 +147,46 @@ def UpdateColumnView(request, column_id, schema_id):
 
 
 def GenerateDataView(request, schema_id):
+    if request.user.is_authenticated:
+        fake = Faker()
+        columns = Column.objects.filter(schema_id=schema_id)
+        column_rows = [{column.name: None for column in columns} for _ in range(10)]
 
-    fake = Faker()
-    columns = Column.objects.filter(schema_id=schema_id)
-    # column_rows = [{column.name: None for column in columns}]
-    column_rows = [{column.name: None for column in columns} for _ in range(10)]
+        for column in columns:
+            for i in range(0, len(column_rows)):
+                if column.type == 'PHONE_NUMBER':
+                    column_rows[i].update({column.name: fake.phone_number()})
+                if column.type == 'FULL_NAME':
+                    column_rows[i].update({column.name: fake.name()})
+                if column.type == 'JOB':
+                    column_rows[i].update({column.name: fake.job()})
+                if column.type == 'EMAIL':
+                    column_rows[i].update({column.name: fake.email()})
+                if column.type == 'COMPANY_NAME':
+                    column_rows[i].update({column.name: fake.company()})
+                if column.type == 'TEXT':
+                    column_rows[i].update({column.name: fake.text()})
+                if column.type == 'INTEGER':
+                    column_rows[i].update({column.name: random.randrange(column.from_range, column.to_range)})
+                if column.type == 'ADDRESS':
+                    column_rows[i].update({column.name: fake.address()})
+                if column.type == 'DATE':
+                    column_rows[i].update({column.name: fake.date()})
+        context = {}
 
-    for column in columns:
-        for i in range(0, len(column_rows)):
-            if column.type == 'PHONE_NUMBER':
-                column_rows[i].update({column.name: fake.phone_number()})
-            if column.type == 'FULL_NAME':
-                column_rows[i].update({column.name: fake.name()})
-            if column.type == 'JOB':
-                column_rows[i].update({column.name: fake.job()})
-            if column.type == 'EMAIL':
-                column_rows[i].update({column.name: fake.email()})
-            if column.type == 'COMPANY_NAME':
-                column_rows[i].update({column.name: fake.company()})
-            if column.type == 'TEXT':
-                column_rows[i].update({column.name: fake.text()})
-            if column.type == 'INTEGER':
-                column_rows[i].update({column.name: random.randrange(column.from_range, column.to_range)})
-            if column.type == 'ADDRESS':
-                column_rows[i].update({column.name: fake.address()})
-            if column.type == 'DATE':
-                column_rows[i].update({column.name: fake.date()})
-    context = {}
+        csv_file = "CSVproject_main/media/data.csv"
 
-    csv_file = "CSVproject_main/media/data.csv"
+        try:
+            with open(csv_file, 'w') as csvfile:
+                writer = csv.DictWriter(csvfile, fieldnames=[column.name for column in columns])
+                writer.writeheader()
+                for data in column_rows:
+                    writer.writerow(data)
+        except IOError:
+            print("I/O error")
 
-    try:
-        with open(csv_file, 'w') as csvfile:
-            writer = csv.DictWriter(csvfile, fieldnames=[column.name for column in columns])
-            writer.writeheader()
-            for data in column_rows:
-                writer.writerow(data)
-    except IOError:
-        print("I/O error")
+        # return render(request=request, template_name="schemas/column.html", context=context)
 
-    # return render(request=request, template_name="schemas/column.html", context=context)
-
-    return HttpResponse(content=context)
+        return HttpResponse(content=context)
+    else:
+        return HttpResponseRedirect("/login/")
